requirements.txt
django
python-dotenv
redis
PyNaCl
numpy
msgpack
cachetools
reportlab
python-docx